feat: split into bin+lib
